Add a template to create an S3 bucket for logging
Initially, this was a template for the bucket, and the associated
CloudTrail trail, until I learned that AWS released the ability to add
a trail for an entire organization in [November](https://aws.amazon.com/about-aws/whats-new/2018/11/aws-cloudtrail-adds-support-for-aws-organizations/)
I reduced the scope of this template to simply define the bucket,
the associated bucket policy, and an archiving lifecycle policy to
move the contents to Glacier.

Using the instructions posted [here](https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-create-and-update-an-organizational-trail-by-using-the-aws-cli.html), I was able to use the bucket
defined by this template in my security account to capture trail data
across all the accounts in the organization.