Add template to deploy a site hosted in S3
Implementing this was a bit more involved than I expected due to the
complexity of configuring the CloudFront distribution, and the
permission dependency on the associated S3 bucket. In addition, instead
of using Let's Encrypt for certificate issuance, I decided on using
AWS Certificate Manager instead.

This wasn't an easy decision, as I'm a supporter of the Let's Encrypt
initiative, and my initial intent was to use the service. However,
I succumbed to the shrinking timeline of finishing this template.

With that said, I would like to deploy a lambda backed custom resource
to retrieve a certificate from Let's Encrypt, and manage the lifecycle
of the certificate.